change js to ts
