+++ @types/NodeJS/index.d.ts
@@ -1,0 +1,5 @@
+declare namespace NodeJS {
+  interface Global {
+    domtoimage: any;
+  }
+}
