+++ @types/global/index.d.ts
@@ -1,0 +1,5 @@
+interface Window {
+  adsbygoogle: any[];
+}
+
+declare var domtoimage: any;
