+++ tsconfig.json
@@ -8,17 +8,19 @@
     "moduleResolution": "node",
     "allowSyntheticDefaultImports": true,
     "noUnusedLocals": true,
-    "noUnusedParameters": true,
+    "noUnusedParameters": false,
     "removeComments": false,
     "preserveConstEnums": true,
     "sourceMap": true,
     "skipLibCheck": true,
     "baseUrl": ".",
     "lib": ["dom", "es2016", "es2017", "es2018", "esnext"],
-    "plugins": [
-      {
-        "name": "typescript-styled-plugin"
-      }
+    "plugins": [{
+      "name": "typescript-styled-plugin"
+    }],
+    "typeRoots": [
+      "node_modules/@types",
+      "@types"
     ]
   }
 }
